requirements.txt
certifi==2020.12.5
cftime==1.4.1
MiniSom==2.2.9
netCDF4==1.5.6
numpy==1.20.3
pandas==1.2.4
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
wrapt==1.12.1
wrf-python==1.3.1
xarray==0.17.0
